requirements.txt
pillow
python-chess
openai
anthropic